Floats split

+++ src/Engine/Events/EventInterpreter.cpp
@@ -386,7 +386,7 @@ int EventInterpreter::executeOneEvent(int step, bool isNpc) {
             startBranchlessDialogue(_eventId, step + 1, EVENT_PressAnyKey);
             return -1;
         case EVENT_SummonItem:
-            SpriteObject::dropItemAt(ir.data.summon_item_descr.sprite, Vec3i(ir.data.summon_item_descr.x, ir.data.summon_item_descr.y, ir.data.summon_item_descr.z),
+            SpriteObject::dropItemAt(ir.data.summon_item_descr.sprite, Vec3f(ir.data.summon_item_descr.x, ir.data.summon_item_descr.y, ir.data.summon_item_descr.z),
                                      ir.data.summon_item_descr.speed, ir.data.summon_item_descr.count, ir.data.summon_item_descr.random_rotate);
             break;
         case EVENT_ForPartyMember:
